refact(#33): util 파일로 분리
