+++ src/utils/index.js
@@ -1,0 +1,9 @@
+import fetcher from "./fetcher";
+import storage from "./storage";
+import urlManager from "./urlManager";
+
+export {
+    fetcher,
+    storage,
+    urlManager
+};
